adding more example code to tools example
