+++ examples/tools/index.tsx
@@ -1,7 +1,7 @@
 import { gsx } from "gensx";
+import { ChatCompletion } from "gensx-openai";
 import { z } from "zod";
 
-// Define a simple echo tool that just returns what you pass in
 const EchoSchema = z.object({
   message: z.string().describe("The message to echo back"),
 });
@@ -18,6 +18,57 @@ const EchoTool = gsx.Tool<typeof EchoSchema, string>({
   },
 });
 
+const GetWeatherSchema = z.object({
+  location: z.string().describe("The city and state or country"),
+});
+
+type GetWeatherProps = z.infer<typeof GetWeatherSchema>;
+
+interface GetWeatherResponse {
+  temperature: number;
+  conditions: string;
+}
+
+// The params will be inferred from GetWeatherSchema
+const getWeatherTool = gsx.Tool<typeof GetWeatherSchema, GetWeatherResponse>({
+  name: "getWeather",
+  description: "Get the current weather for a location",
+  schema: GetWeatherSchema,
+  function: async (params: GetWeatherProps) => {
+    // params.location is now properly typed as string
+    console.log("Getting weather for:", params);
+    return {
+      temperature: 72,
+      conditions: "sunny",
+    };
+  },
+});
+
+interface ToolExampleProps {
+  message: string;
+}
+
+const BasicToolResponseExample = gsx.Component<
+  ToolExampleProps,
+  GetWeatherResponse | string
+>("BasicToolResponseExample", (props) => {
+  return (
+    <ChatCompletion
+      model="gpt-4o-mini"
+      tools={[getWeatherTool]}
+      messages={[
+        {
+          role: "system",
+          content:
+            "You are a helpful weather assistant. Use the getWeather tool when asked about weather.",
+        },
+        { role: "user", content: props.message },
+      ]}
+      //gsxExecuteTools={true}
+    />
+  );
+});
+
 async function main() {
   console.log("\n🔊 Testing echo tool...");
 
@@ -28,6 +79,18 @@ async function main() {
   // Test JSX usage
   const result2 = await gsx.execute(<EchoTool message="Hello JSX!" />);
   console.log("JSX call result:", result2);
+
+  // Test a tool with a chat completion with a tool response
+  const chatResult = await gsx.execute(
+    <BasicToolResponseExample message="What is the weather in San Francisco?" />,
+  );
+  console.log("Chat completion result:", chatResult);
+
+  // Test a tool with a chat completion with a tool response
+  const chatResult2 = await gsx.execute(
+    <BasicToolResponseExample message="Hello" />,
+  );
+  console.log("Chat completion result:", chatResult2);
 
   // Log the OpenAPI schema
   //console.log("\nTool schema:", EchoTool.getOpenApiSchema());
